requirements.txt
python-telegram-bot==21.6
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
rapidfuzz==3.10.1
cachetools==5.5.0
google-genai==1.3.0
